Extract importer accumulator TypedDicts into importers/common
Create a new `importers/common` package and move the three per-position
accumulator TypedDicts (SecurityPositionData, CashPositionData,
SecurityNameMetadata) there.  These were defined inside ibkr_importer,
which caused newer importers to reach into the IBKR module to reuse
them.  Nothing else moves in this commit; IBKR just imports the types
from their new home.

No behaviour change.
